Update and rename 27. remove element to 27. Remove Element
Approach 1: Two Pointers

When iterating element in the array equals to the given value, skip this element. As long as iterating element is not equal to the given value, we copy an iterating element to the place of where the given value was found, and increment both indexes at the same time. Repeat the process until the array reaches the end of the array.

**Time Complexity**: O(n).
**Space Complexity**: O(1).

Approach 2: when elements to remove are rare

Useful when there is few elements to remove, f.e. we remove element 4 from the array like [1,2,3,5,4].
When we encounter nums[i] = valnums[i]=val, we can swap the current element out with the last element and dispose the last one. This essentially reduces the array's size by 1.

**Time complexity**: O(n). 
**Space complexity**: O(1).