Update 58. Length of Last Word
Approach 1: **Iterate String from index 0**

Iterating the string from left to right. There are some problems with spaces after the last word: we can solve it with  trimming the spaces before or at the end of the string. If modifying the giving string is restricted, we can create a copy of the string. 

Approach 2:  **Iterate the string from the last index.**

The idea is to start incrementing the count when you encounter the first alphabet from the last and stop when you encounter a space after those alphabets.

Approach 3: **Using split() and list**

1. We have to split the sentence by spaces using split().
2. We split all the words by spaces and store them in a list.
3. Print the length of the last word of the list.