Update 83. Remove Duplicates from Sorted List
Approach:
Traverse the list from the head (or start) node. While traversing, compare each node with its next node. If the data of the next node is the same as the current node then delete the next node. Before we delete a node, we need to store the next pointer of the node 

Time Complexity: O(n) where n is the number of nodes in the given linked list.

Another Approach: 
Create a pointer that will point towards the first occurrence of every element and another pointer temp which will iterate to every element and when the value of the previous pointer is not equal to the temp pointer, we will set the pointer of the previous pointer to the first occurrence of another node.

Time Complexity: O(Number of Nodes)
Space Complexity: O(Number of Nodes)