Create 88. Merge Sorted Array.py
https://www.tutorialcup.com/leetcode-solutions/merge-sorted-arrays-leetcode-solution.htm

Approach (Sorting):
We can transfer all the elements of the second array to the first one. We can then sort the first array to get the required result.

Time Complexity
O(KlogK), where K = N + M. N = size of the first array, M = size of the second array. As we sort the first array after it has stored all N + M elements.
Space Complexity
O(1) as constant memory is used for variables.

Approach(Two Pointers)
We can use two-pointer technique to merge the two sorted arrays, into a new array. But, the creation of this new array will cost extra space. We can try to avoid this extra array especially when the first array has enough space to hold all the elements of both the arrays. We can use two pointers and start merging the elements in the back of the first array.
This will cut the problem of “extra array memory” as we keep fixing the elements in the void space.

Time Complexity
O(N + M). N = size of the first array, M = size of the second array. As we traverse both the arrays once.
Space Complexity
O(1), as we store all the elements in the first array.